backend/requirements.txt
# ============================================================
# 🌾 AgroForgeSIM Backend - Runtime Requirements (Production)
# Python: >=3.10
# ============================================================

# --- Core Web Framework ---
fastapi==0.115.2
uvicorn[standard]==0.30.6

# --- Data Validation & Settings ---
pydantic==2.9.2
pydantic-settings==2.5.2
typing-extensions==4.12.2

# --- Config / Environment ---
python-dotenv==1.0.1

# --- HTTP & Async I/O ---
httpx==0.27.2
aiofiles==23.2.1

# --- Data / Engine Utilities ---
pandas==2.2.3
pyyaml==6.0.2

# --- CLI ---
typer==0.12.5

fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
